refactored webviews, register and notification listeners
