--- vscode/src/configurations/listener.ts
@@ -1,34 +1,0 @@
-/*
-  Copyright (c) 2023-2024, Oracle and/or its affiliates.
-
-  Licensed under the Apache License, Version 2.0 (the "License");
-  you may not use this file except in compliance with the License.
-  You may obtain a copy of the License at
-
-     https://www.apache.org/licenses/LICENSE-2.0
-
-  Unless required by applicable law or agreed to in writing, software
-  distributed under the License is distributed on an "AS IS" BASIS,
-  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  See the License for the specific language governing permissions and
-  limitations under the License.
-*/
-import { ConfigurationChangeEvent, workspace } from "vscode";
-import { userConfigsListened } from "./configuration";
-import { globalVars } from "../extension";
-
-let timeout: NodeJS.Timeout | undefined = undefined;
-export const configChangeListener = (params: ConfigurationChangeEvent) => {
-        if (timeout) {
-            return;
-        }
-        timeout = setTimeout(() => {
-            timeout = undefined;
-            userConfigsListened.forEach((config: string) => {
-                const doesAffect = params.affectsConfiguration(config);
-                if(doesAffect){
-                    globalVars.clientPromise.restartExtension(globalVars.nbProcessManager, true);
-                }
-            })
-        }, 0);
-}
