+++ vscode/src/lsp/notifications/register.ts
@@ -13,17 +13,13 @@
   See the License for the specific language governing permissions and
   limitations under the License.
 */
-import { JdkDownloaderView } from "../webviews/jdkDownloader/view";
-import { extCommands } from "./commands";
-import { ICommand } from "./types";
+import { NbLanguageClient } from "../nbLanguageClient"
+import { notificationOrRequestListenerType } from "../types"
+import { notificationListeners } from "./handlers"
 
-const invokeDownloadJdkWebview = async () => {
-    const jdkDownloaderView = new JdkDownloaderView();
-    jdkDownloaderView.createView();
-}
-
-
-export const registerWebviewCommands: ICommand[] = [{
-    command: extCommands.downloadJdk,
-    handler: invokeDownloadJdkWebview
-}];
+export const registerNotificationListeners = (client: NbLanguageClient) => {
+    notificationListeners.forEach((listener: notificationOrRequestListenerType) => {
+        const { type, handler } = listener;
+        client.onNotification(type, handler);
+    })
+}
