+++ vscode/src/webviews/jdkDownloader/prompt.ts
@@ -14,7 +14,7 @@
   limitations under the License.
 */
 import { commands, window } from "vscode";
-import { l10n } from "../localiser";
+import { l10n } from "../../localiser";
 
 export const jdkDownloaderPrompt = () => {
     const downloadAndSetupActionLabel = l10n.value("jdk.extension.lspServer.label.downloadAndSetup")
