+++ vscode/src/webviews/jdkDownloader/styles.ts
